added norm scaling / delays.py to handle delays
